requirements.txt
requests
linkedin-api  # https://github.com/tomquirk/linkedin-api
python-dotenv